fix(network-manager): run as a service if systemd module is present
In the current state, services that depends on network need to
use dracut hooks, since nothing with pull in the network
targets into the transaction.

In the future it would be nice to provide developers on systemd-only
systems the possibility to not use dracut hooks at all, but simply put
normal systemd services into the initrd.

Also some modules even right now depends on systemd ordering, like
cryptsetup, so let's make sure, that the ordering inside systemd work
properly as well.